requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.28
python-dateutil==2.8.2
gspread==5.12.0
google-auth==2.23.4
reportlab==4.0.4
Pillow==10.0.0 